These are the nav package pieces I use + some bonus parts
A little bit of browsing has its benefits. All parts have a strength factor that is never the problem and a bolt count that might hurt a bit. Beware that these parts came off of ships I use regularly so there may be references to parts not here. LMK if you find those. Read the chips.