Added example levers and updated the brainbox
Memory relays as chip sockets is basic tech these days and deserves to go on the basic version. Everyone can make lighter things now :) Do not that this does not make use of MemSocket80 so the advanced version (in extras) still has cooler toys.

This version also has a basic lever bar and sanner with all the fields pre-named. Why tell you to make them when I could just give them to you?