Replaced the scanner with AASStop as core suite
AASStop is well proven by now and should most definitely join the core package. It will replace the scanner script to maintain a small footprint.

The scanner didn't even have a panel by default and isn't necessary for navigation, which is the name of the package. It's going in extras.