Redudnant parens(Yolol order of operations WHY)
This version won't notify you if it is avoiding, requires separate laser control, and needs to be turned on/off via chipwait -1 shenanigans, but it is twice as fast. Designed to be used with the AAS_Stop1LCompanion chip that does all of those things. Initial testing is promising and the change to the algorithm can easily save a bit of space in the original AAS_Strafe.