Line error out logic
The early /0 line cancel prevents momentary false activation during shutoff. Costs a rename of :AAS and that hurts. Also doesn't notify you if you are avoiding, but that gets moved to the companion chip AAS_Stop. You may not like it, but this is what peak performance looks like :p