-redundant control flow logic.+collission export
The added control flow was dumb and unnecessary. The shortening of the :AAS field to :AS made room for :c field export again! Still a pain to run as you have to change all instances of AAS to AS in the rest of the code bundle, but if I like this setup Ill make all those changes for people.