Added copy pase features
You can now import a waypoint and grab its string on the import button to copy-paste into any text comms and other people can load up the coords by copy-pasting that string into :WEC in the export button. Also converted button styles 1 so that they pop out when the operation is complete. There are some changes that go on in WPIncrement as well, but I do not intend to make them in the repository here because this chip is not part of the basic bundle. This and the WPIncrement changes are going in the advanced bundle in a few minutes with some much more notable changes.