Update broke code. Fixed
It seems that the mat field does not change the moment the index field is changed anymore. When it changes is based on a metric of eldritch origin. New code just waits until the material value changes and then keeps going. It does mean that only one material will make the system cry, but if you just cancel the scan it will give you its data so far.

It's ugly and not optimized but it will work for now. Ill put more work into it later