More Scanner change shenanigans
Added more waiting before the scan goes thorough because large roids seem to add their masses up in a piecemeal fashion. I'm not sure what the exact timing on line 6 should be, but I imagine its network related so...tune it till it works. Three ticks worked for me but no wait lead to a great many confusions. Also the goto on line 8 totally skipped line 9 and that is not good. Probably a result copying these line by line from in-game chips without a copy function and missing a line.