Automated Control Test
This is a copy of the code cluster and levers modified such that the chips do not control the lever values directly. Rather, they control the LeverCenterOutput and let the levers interpolate to the desired value. This change is has to intentions: it will help resolve issue that arrive from instant inputs to levers(un-commanded extra-axial movements) and it will also allow players to override inputs while they are currently operating (as it simply changes default values rather than possible inputs.)

Normally I just update the advanced version as it is my playground, but this change is fairly major and deserves some review. I am very hopeful for this change and invite feedback and comparison to the previous version.