Lever Centering Control Update
NavSuite_Adv now uses the LeverCenterOutput fields where applicable. This creates a smoother movement and removes all of that ugly instantaneous pitch/yaw from overthrust during acceleration. As an arguably bigger upside, this means that the controls can be overridden at any time by the player with zero interference from the computer so long as the controls are being manipulated by the player.

This update also comes with a few QOL upgrades such as:
Clean, bolt-less dashboard
Clearer dash naming(in select places)
More example pieces
A large tank version of the refuel point

I hope you guys like the new changes as much as the testers. As always the old version is in the Old folder.
In the future I will be working in ways to remove direct lever control from a few of the remaining auxiliary systems the use it. I will be updating the example ship to be heat ready and include this new update. And finally, I will be updating the basic module with these features where applicable.