Update ExampleShip
Example ship now uses NavSuite_ADV6.

This is an updated version of the original system integration lab that I tested the first NavSuite on. It is small and cheap and still comes with all the bells and whistles. Easy transport from A to B if you like or a copy of the working system(minus the scanner and CLF speedometer, but you can put them on as the code is in there) to disassemble for learning.