GalacticAlign_LeverCenter
As a replacement to NavSuite_ADV6's GalacticAlign that does not use LeverState directly. I am working on creating replacements for the rest of the systems such that naming can be arbitrary. Current limitations in math understanding leave me concerned about Approach modifications and the impact of lever centering delays. I will consult smarter people in the coming days.