LeverCentering Approach
After some consultation I was surprised to learn that PIDs are generally modeled such that a delay is accounted for within the math. As such it is a simple matter of changing the controls from direct LeverState to LeverCenterValue. This was the last bastion of necessary LeverState Control. As it stands now, you can name your levers whatever you like and the 2 letter standard can move off of the player controls.