Passive controls update for the basic version
This basic NavSuite update moves the levers over to the passive control system, basic binds, and adds some example pointing lasers. Players were confused about how to use the built in approach features, these pointing lasers (being on when AAS is off) should serve as an example of how to set up your laser array such that approach will work. No approval period for this one, the legwork was already done for the advanced set. As always contact me if you can't figure it out and the wiki doesn't help.