Larger CRUS window and emergency fall-through fix
Little bit of reordering. Autogen is back to line 3 again so that if buttons get shot off and you didn't build backups, the emergency case activates. It has been set after the Crus-AAS linkage so that it doesn't break those if buttons go missing. Also set the :CRUS comparison to less than 0.001(smallest positive yolol number) to account for ships that want a :CRUS value between 0 and 1. One could do the same for Backwards and b=99.999 but I'm not going to at this time. Feel free to add those 4 chars to line 1 yourself if it is super important to you. Also changed the comment on line 18 to reflect the new fuelchamber name shorten. This is an easy, drop-in replacement. Fixes will go into an ADV7 update soon.