NavSuite_Adv7
-Minor update to 7: GenericStates chip tweak(drop-in replacement) and fixed an ugly wire in the CLock speedometer. Nothing major. Original commit message follows.

This version finishes off the conversion of the auxiliary systems to the passive control regime. The navsuite is now LeverState/bind agnostic. If you want to use anything other than default binds, all you have to do is change the names on lines 2 and 3 of the state checks in the GenericStates chip.

Also of note: I am no longer hiding the advanced version away in extras. I was originally worried about confusing the basic endos, but it is my hope that the labeling and (upcoming) limited options make it simple enough. Coming up: I intend to move all the code into its own folder out of main and convert them to their advanced versions. It is my estimation that the people who intend to dig through the code are probably looking for the advanced documentation anyhow.