Update Scanning.yolol
Put line 4 into line 5. Makes it coniuously try to scan which is slower, but also still functional if you lose the rock and re-aim at it. After the alstel trip, this seems like the way to go. Technically this means line 4 doesn't need to exist, and that update will come with NavSuite_ADV8 The rest of today's changes will be the same. Update your chips if you like, but they are just convenience updates.