Update LaserPanic+SpeedCLock.yolol
Added optional upgrade that keeps it on in the event that the CLF turns off despite the state telling you it is on. Requires a rename to keep it the same number of lines. You could also take the last [if then end] area and just put it in a line below without renaming things on your ship, but then the speedometer would update every .6s and you would have to set up the :LC+= clause on that line too if you want it to start lasers charging as fast as possible when you hit fire