Poking about the Speedo cage reliability option
It seems that, while setting the cargo lock off and on on the same yolol line fixed the speedometer yesterday, today it errored out and did not reset. Splitting the Cargo lock reset over two lines may sort this out, looks good for now, but so did the last one.