Example ship with working slip sensor
The other .fbe is hard to work with because that you need to re-calibrate the laser due to the fact that it was moved(needs to be done every time it seems). This one just has it pre-calibrated and and on a ship so you can see it in action.