LDVectors Chips WIP
Here are the yolol files for the new laser designator translation vectors modules. These output ship speeds on all three axes relative to the ship. The slope constant is set up for 30FPS and with Vsync off. All still WIP, but the numbers are good now and it should be much more robust for use in Strafe Arrest/Emegency Stop/Gravity Cancel systems. I still need a volunteer with a strong PC to make the 60/90/120 versions of the constant. Contact me if you'd like to help.