Updated Laser Designator Translation Sensor
This version is a little bit cleaner when it comes to small sensor lengths. I set sensor distance from the laser at a standstill to 6 for the smallest lateral speed chunks, but I'm still not convinced its best. Have a poke at it and see how you feel. I also still need testers for the 60/90/120 FPS versions