Azurethi's Rework
A big thank you to Azurethi for essentially rewriting this whole system and actually get it working the way I expect it to. It turns out that I did not understand rotation matrices well enough to notice that the gyro uses a non-standard rotation order. I still don't understand most of the mechanics of this math, but I was able to convert the result into NavSuite compatible format. Inertial navigation is making progress. The project can now move forward and the next step is a chip that takes our relative motion and converts it into absolute motion using the rotation matrix generated here. There are a few things I want to clean up in the main project too. Line 7 of WPImportExport missing a few chars on the end of line 7(not sure what happened there) Ill try to get as much of this ready for EosCon as I can.