Speed Modulated Laser Length
This update continuously updates the laser designator length as pat of an effort to increase the granularity of the lateral speed data. This version now has a granularity of 0.013 at low speeds versus the previous version's 0.180.  Thank you to DarkyShadow for pointing out that it needed to be improved and I hope to see your addition to this solution soon.