Speed Modulated Laser Length
This is the downloadable for the next iteration of the laser designator speedometer. Developed in conjunction with DarkyShadow, this version has a lateral speed granularity of 0.013 at low speeds versus the previous version's 0.180. This is done by setting the laser designator length to a function of the ingame speedometer device such that the length scales by its total possible deflection at any given speed. This tool is now better suited to precision operations. It is still only built for 30FPS as I have had nobody come forward to provide data for higher FPS constants. If you wish to contribute to this project or any other part of the NavSuite, please contact me.