Missed a signflip
As part of the adaptation of Azurethi's gyro solution to NavSuite and ISAN axis scheme, I flipped the X axis to be left positive and right negative. Unfortunately, I missed a sign and it meant that this was not always correct for all orientations of forward motion on the X axis. Hopefully this is the last error that arises from this, but this kind of thing is why the systems stay in WIP to bake for a bit.