Line Que LDTranslationSensor + Cosine
This version of the code uses Firestarr99's Line Execution Queuing technique introduced at EosCon. This means that it no longer runs asynchronously and still runs at 0.2s, but costs 5 chips. It's like chip execution queuing, but is more stable if multiple ships are hosted. This version takes 5 chips all of which operate this same code so make 5 copies if you try to use this. It also uses 2 extra globals that you need memory field for (:A and :SO) as they don't exist on the sensor. Ill make a module with all 5 in there if this ends up being the production version..

I removed a redundant ABS function. I don't know how it got in there, but by the end of making the last version it did nothing.

This version also implements Darkyshadow's Cosine suggestion. Sine and Tangent function in yolol can't output values for anything below sin 0.058. Ostensibly that is also true for cosine(90-angle) but it still seems to produce more stable results. This may go the way of the ABS function later on.