0.2s Gyro Matrix
For the not so low price of 6 professional chips and 2250kg of misc support crap you can have your very own line execution queuing gyro!
This version is insulated from the rest of the ship because we use almost a full alphabet's worth of globals to get things done. If you use this only wire into the very ends on either side your your ship and the module will cry. Takes in: 1 gyro. Puts out: 9 variable rotation matrix and an unused reference to gyro calibrate every yolol tick. Go ahead and remove that GC reference if you want, I just don't want to grab the file again and it doesn't hurt anything.