Re-arranging some variables
This update moves some variables around for redundancy and ease of use. A copy of all internal variables is now in the memsocket so that you can remove either or both internal side memory chips and the code will be just fine when you put them back in. I also exposed W on the internal side input chip so that you can do reset things with it and removed the other garbage littering that chip.

This is a QoL update. If anyone is asking "Do I need to replace this on my ship?" the answer is no. No code has changed and the only difference is that the new one handles memory chip removal better and has the ability to reset the chips externally.