NavSuite 8 Release Candidate 1
I am going to put out a release before inertial navigation as there are still kinks to iron out and there have been changes internally that have never made it back into the system.

This version includes a human input detection system. Many people disliked that you had to hold reverse to turn off the automated systems. This HID system allows players to turn off automated movement through any normal flight keys. If you want a key to not contribute to that, simply change the output variable for that control to something else.

The chair has been updated to be lighter by redoing the beaming and ducting. It is now under 5000kg and includes all the same features. I also moved the levers to be part of the chair package and is included in that weight. Could be a bit lighter still, but I am happy with where it is.

I have moved the receivers off of the single character list. The compass system does not need them for brevity and it frees them up for use elsewhere. I moved them to the SR+FBD prefix for Sensor.Receiver.

The LDTS replaces the Moriwatari CLF speedometer in this version and outputs relative speeds on three axes. It will use a gyro later and combined they will do inertial navigation, for now this part is pretty good and going in. I am not sure if flipping it over to hardpoint on the bottom like it is here is going to change the vectors, but if it does after I use the package, Ill update the code to flip the signs. In the coming release candidates I'll include Strafe Arrest and ESTOP functions, for now I am working on making them play nice with AAS.

LookAtTheBunny pointed out that a fix which was in the approach script on the github for ages has been left out of the official blueprint. Now you can turn off approach while it is in operation and have it reset the levercenter values in stead of leaving them where they were.

The KR value in compass now sets to -20 in stead of 20 and the chips it affects have had their relevant sign flipped. I intend to use this in resetting the 0.2s LEQGyro later.

I moved some of the optional and example pieces off behind the main cluster. These parts can be included in the ship (like the refuel points) but are mostly there to give examples of how things can be done. In here I have included radiator placement on things like generators and heat-sinks and I am open to opinions and suggestions on this regime. I may just make a separate building techniques blueprint in extras at some point.