Range Calulator Chip
HappyTrigger's autonomy range calculator algorithm minified onto a single chip(well, some of it). Thank you to HappyTrigger for walking me through your system so that we could bring this down to size.

Runs on averages over a variable number of cycles for smoothing. Higher Z values make the output smoother at the cost of a longer stabilization period.

May eventually replace the DTW panel as it's data is often less relevant and/or covered by other instruments.