NavSuite_Adv8 Release Candidate 2
Just a quick re-upload of the file as I forgot to bolt in the FCUs in the rebuilt brain cluster. Not going to update the version for such a small(but important) change.

This is the first integration of the inertial navigation system. I am using the 0.6 single chip version as it is lighter. If you want the 0.2s cycle module there are instructions for how to use it and it will be in Extras on release.

This version also includes strafe arrest and emergency stop buttons. These features mesh with AAS to allow avoidance, but limit strafing into another asteroid laterally. They do include a PD controller so...tune that.

There is now an autonomy calculation where the old DTW panel would be. It was one of the most commonly added features so now its in the suite. This version is based on HappyTrigger's algorithm condensed into a single chip.

There is also a human input detection device included that can be sat near the seat. This device detects if the levers are being controlled by humans specifically. Current iteration cancels all automated movement if human input is received. Currently it just sets the levercentervalues to 0 but you could also make it turn everything off if you like and select or split out individual controls.

I have also cut out some extraneous stuff like the lasers and associated script. They will show up again in the Extras folder. Core example pieces like the transponder, fuel chamber, batteries, and tank supports have stayed. I built a test ship out of these parts, but there are still more test to be done. If anyone else wants to help verify functions I would appreciate it.