Minor waypoints tweaking
In taking a break from trying to make a one line version of the LDTS that doesn't care about your framerate, I poked around Archaegeos waypoint system.

This version tells you if the waypoint is still loading directly on the WP screen and now updates immediately when you save and manually enter waypoints. Not a big change, but it should make it much easier for new users and makes me feel better about dropping the DTW screen from the included chair. 

You can still access the :DTW text panel string and do whatever you like with it, I just found that I only ever looked at that text panel when I needed to know if things were still loading and that's a dumb reason to keep a panel around.