LDTS update and cleanup
The folder was a mess and out of date

This update includes Azurethi's shortened RZ assignment and is thus able to fit the SL assignment on the same line. The SL assignment has been moved to LDTSForward to reduce error on the forward vector when moving laterally.

I have adjusted the m constant to be a little higher. This seems to provide a little bit more parity between the speedometer value and the amalgamated axes on average. This should improve INS accuracy slightly.

I removed the extra .fbe with the inferior conficuration. There is not only one .fbe file and it comes with all the changes to the code. It is not the exact same assembly as the last set because the last one didn't want to move a lot of the time. This one is spaced a fraction of a hair farther apart and that allows it to be tuned more easily. As a byproduct, the b constant was slightly increased, but it does not meaningfully affect performance other than the code and device are not interchangeable piecewise without some minor constant edits.

Anywho, its better and the area it lives in is cleaner. Next up is the quest for a framerate independent version.