Yet another LDTS sensor rebuild
Sinsidious mentioned that they were having trouble moving the sensor and designator pieces around again. Turns out that the infamous 89.789456==90.000 rotation bug had ruined the whole part. I thought that those tiny errors were just some quaternion rounding shenanigans that were cosmetic only but they are most certainly not and I hate it. Apologies for those affected, this one should alleviate that issue. As always: feel free to message me with any issues.

I also tool this opportunity to snap with hardpoints to a perfect 120cm apart which did change b again, code is in there and I have edited the constant on the yolol file in this folder to reflect the new part.