NavSuite_Adv8
This version includes Inertial Navigation, XYZSpeed, Strafe Arrest, and ESTOP on top of the previous features.

The Inertial Navigation System(INS) provides arbitrary coordinates outside of the ISAN bubble. You can use this to make capital ship relative coordinates around moons. It will lose accuracy over running time/distance just like a real INS so runtime is an important concern, but it seems to be sufficient to get to zone 3 of a moon belt and back to capital ship. The navigation systems have been modified to use either INS or ISAN mode. There will be a larger and more advanced version in the Extras folder in the near future with the goal of it being marginally more accurate.

The MoriWatari CLF speedometer has been officially replaced with the Laser Designator Translation Sensor(LDTS). This is a speedometer that reports ship relative speeds on the forward, horizontal, and vertical axes. It is also smaller and lighter than the requires size of CLF speedometer to do all speeds on all axes.

Using the LDTS we can do things like Emergency stop and Strafe Arrest to limit un-commanded movement in say...a moons gravity well. The AAS has been upgraded to now stop drifting laterally after the ship has avoided the rock if your turn on Strafe Arrest. ESTOP will attempt to stop the ship regardless of its orientation/motion on any axis.

There is a new panel on the left side of the cockpit chair: EstRange. This panel will give you a range estimate for your ship based on your current resource usage and speed. This was one of the most commonly added components to the chair so I felt like including it.

Adding the above panel cost the DTW panel. My primary usage of the DTW panel was to see if waypoints were loading. As such, I have moved the loading indicator onto the WP panel.

There is also a new system for detecting the difference between human input and computer inputs. If you place the HID near the pilot chair, the ship will put out a non-zero value on :HI to denote when there is human input.

There could be a bunch of other small changes in this update but these are the main things that I can think of right now. If you have any questions, I am always here. Documentation will be on the wiki as I get the motivation to do so.