Adv8 Example ship
Somebody asked for an example of a working ship using the latest version. This is the test shuttle I built just before 8 came out. Every time I release, I build a little ship with all the parts just to make sure it works out of the box. If there is no example ship for the current version in the future, feel free to ask for that version's system integration lab. If you have a ship you would like to be included in the examples, that would be super cool.