NTPS' Framerate Independant LDTS
NTPS has been working on a version of the translation sensor that does not require a constant frame rate. We sat down the other day and ironed out a few bugs. This is the most stable version we came up with bar a disagreement over the ideal laser length scaling math(If you read these NTPS, the latest version exacerbates the Z axis issue immensely, it seems best to leave S where it was). It is a LEQ chipset requiring 5 copies of the chip and network isolation/aliasing via memory relays. Ill put up a copy of the whole working yolol set and device for people to play with soonTM. There are still some things to work out. For those who want to play around with it now, the comments should make this somewhat less inscrutable.