Frameless LDTS Demo
A small mockup of the code and parts for the framerate independent LDTS. This version comes with a more precisely calibrated sensor module(0,0 @L=+0.16) that should facilitate more accuracy in all versions. I will update the internal copy of the NavSuite with the new part so that it comes with the next version. I also swapped one of the utility bodies to its mirror for a more even footprint. I am not sure how it got to be the way it is in the ADV_8 bp as earlier versions were the way this one is now and that is what is on my ship.