Frameless LDTS Demo V2
Now with an actually precisely calibrated sensor. This thing stays 0,0 at any laser length. Turns out that snapping to perfect 90s is not prefect alignment for the laser. I had to yaw the sensor by 0.0004 degrees and then re-align to make it perfect at all lengths. "Why is this a thing" you may ask. All I can offer you is my sincerest condolences and this device.