NavSuite_Adv9 Candidate One
This version uses the new Frame-less LDTS code and zeroed sensor at all lengths.

The brain cluster was reworked to fit the LEQ system and in doing do I flipped it over such that the relays are now on the bottom and the FCU/MFC on top. This way you can put it right behind the cockpit chair and they will snap together. It also provides slightly easier access to the FCUs. Slightly higher SF for fewer bolts, but the total weight is  ~25% higher as it holds more chips (Frame-less LDTS is like 10 chips versus the previous version's 2)