NavSuite Adv9 Test Ship
As always, I throw together a shuttle with the candidate file to make sure they work out of the box and not just on my ships what may have legacy crap. Caught a few things this time relating to the new chair and LDTS chipset. Also strafe was keeping you drifting in stead of stopping you.