Galactic Align cleanup
This rework stems from tuning a bigger ship and just snowballed into a bunch of "yeah I should fix that..." A similar thing will be happening to the FixerIDNav chip soon as the way this change handles rotation is preferable to the current operation and I can already see a bunch of stuff I don't like in that chip. As always, both will sit in WIP for a bit to bake before I put them into a navsuite iteration.

Moved the inactive wait loop to a non assigning line. This is some minor performance paranoia, probably doesn't really matter that much, but it is cleaner.
Compacted the pitch and yaw lines into one. This makes thruster timing better and would ostensibly increase response times if not for the mandatory wait period for the compass/rotation matrix to update.
Added timing variables to the roll and pitch/yaw lines for variable burn times per operation which is useful for larger ships. This is also tuned for a large ship, so don't use it as is unless you also have 1000+ crates. A tuning guide should be in the wiki and if not it basically follows the same pattern as the Nav chip tuning.
Fixed a bad goto on the missing button catch. TBH the fact that it went to a blank line is a bit embarrassing but its been that way for all to see so...yeah. Now it isn't.