Update FixerIDNav_modded.yolol
I have been meaning to clean this one up for a while. TLDR its shorter and slightly better. Still WIP so the usual "at your own risk" applies but all tests check out so far.

Changes:
Removed line 3 and shifted all lines up. I shoved its behavior onto line 4 of the new chip and also made it turn on strafe arrest.
Traded out a couple of sqrt operations for ^.5
Removed a lot of unnecessary whitespace in "if then end" operations and removed some altogether (such as the old line 3)
Replaced the velocity calculation with a speedometer read.
Merged lines 6 and 7 as they had room after the velocity calc was removed.
Split apart the orientation checks from line 11 and re8did their comparisons on the following line so that the turning lines could be shorter.
Merged lines 13-14 so that pitching and yawing happen on the same tick. Rolling is now done only if the ship did not move last iteration. Moved some behavior off of line 12 and put it on its goto at line 14. That line is still a bit of an odd jump, but its better than if it were next in line.
Changed the chip name. It seems odd and unfair to all the other contributors that their name is on the chipwait in stead of in the comments, also it looks almost nothing like the original

Possible futures:
I may want to add in thrust timing and gotos on 11 and 12 so that it supports longer and shorter burns. Right now its always a 2 tick burn and that seems like its sufficient so this one is up in the air.
Possibly remove the "if then end" from line 12. It would probably be sorter without it, but id also have to re-do the comparison so that it can only output 0 or 1 in stead of 0 and non-0. In the end it really doesn't matter right now as I don't need the space, but I want to note it down in case I want to do it later.
I could also rework some variables to be 1 char shorter and that should save 4-10 chars per line. That may reduce the number of required lines by a few. This seems like it would be worth exploring at a later date, but I am le-tired and am going to salvage what I can of my sleep schedule.