Large gens bad. Use these instead
In an effort to spread awareness about the inefficiencies of large generators I put forth this blueprint. Inside are 4 arrays that handedly beat large generators in energy/mass and volume. Two sets of any of these configurations produces 108ke/s(1.5x the power) and use ~1/3 of the mass of a large gen/processor/tank set. Each module is made up of 4 self contained generator sets that do not need external cooling. You can further split that up to 8 if you remove the connecting heatsinks and place a few around your ship in stead. Feel free to piece these apart or stack more of them together as you see fit.