Faster Orientation and Nav Changes
Minor tweaks here and there to make things faster. Its been a while since the last version and that time has given me some fresh eyes to stare at this with.

- Compass Orient is now a 3(was 5 ticks) tick operation and does not calculate Lxyz
- Compass DisplayPrep now handles Lxyz with the same 5 tick frequency that orient used to do it in. Looking at it now I may shift the whole line 3-7 loop down by a line for sync reasons.
- Nav is again a bit shorter and reordered to make the best use of the time. Timing variables have also been added for more tuning options.
- Galactic align should be largely unchanged since the update last week, but I figured that committing the whole WIP folder at once made sense here as all these pieces go together in the clean-up and speed-up drive.

Next up: 
Ill be making a small change to the MovementArrest system(Several people have been calling it inertial dampening now and that's a fitting name) such that it remembers your strafe arrest setting before ESTOP and restores it to that setting afterward. Should be a minor change as there is room and it doesn't really affect its active function. 
There is also a change that I have been sitting on for a very long time suggested by AdrianRC: an compass ISAN3in mod that waits if the receiver hasn't updated. I'm still wary about touching such a core system and generally haven't noticed a need for the change(receivers usually update in about 1.3 ticks) but if there are people who have noticed the issue please do come show me.
Finally, when this change goes through id like to take a look at the basic version as it has been forever since I have and it could probably use some love.