Test Integration
The next release is right around the corner and this is how I start the approval process. This one is a general improvement pass focusing on QOL in installation and general  tweaks throughout the chips to speed them up and improve their workings. Test ship is built so that I can confirm that everything works out of the box, then I integrate it into a big boy ship for long term testing.

Notable changes: 
New chair, new sensor cluster, new brain cluster. Same stuff in total but fewer individual parts to slap on the ship. Should lead to fewer errors in assembly. All the usual modules still exist though so advanced builders can take things apart and put them where they see fit.
Generic states has been changed to always use the Human Input Detection system, and that part has been integrated into the chair.
The brain cluster now has a ship diagnostic scanner in it. I am not sure what I need it for, but there was room and I didn't want to put the FCUs under the MFCs as that would make them hard to tinker with.
Strafe arrest now remembers its state after you press the ESTOP button. I find that I have been using the ESTOP as a general brake and I want whatever I set my Strafe arrest to before hand to stay that way after I'm done with the stop.
Activation delay on the LDTS chip responsible for naming the Laser Desigantor-Sensor combo. This should allow the use of the :transponderOwner field to personalize the :SI names and better avoid nearby ship problems out of the box. You may still want to change them as it is easy for onlookers to figure out what the value is, but that was even more true before when it was always the same if you didn't change it. At least they are now unique per person so that multiple users that didn't do any personalization can play side by side.
Intertial Dampening has had some naming convention changes. Those have been propagated out to all affected systems.
Nav has been renamed Autopilot and has received an overhaul. It has faster loops, more tunable turning burns, and removed a lot of ifthenelseend ugliness.
The rotation matrix now calculates orientation every .6s on the single ship version rather than the 1s loop it used to be. Lxyz is now calculated over in CompassDisplayPrep.
Galacctic alighn has had the same orientation improvements made to it that Autopilot has.
I have also fixed a small cable that was not part of the chair module causing chairs to mysteriously not work when people moved them by the module. I meant to fix this ages ago...but life got in the way.

As always feel free to download the package and the test ship for your own testing and to provide any feedback. Every single new player I have seen has gone straight for the advanced stuff. The more eyes we get on these candidates the more likely they will be easy for new players to put on.