Fixed some test release things last week
Life has been a little bit hectic recently so I have had this for a week or so but not pushed it out.  TBH I can't remember what the specific errors I fixed were, but somebody is trying to test it out so I am updating the test kit and vehicle before work and I'll come back and do the individual chips after to see what changed in each. I lost my physical notes in moving so documentation here may be a bit spotty while I stabilize.