Adv9 SIL
This is the system integration lab for version 9 of the advanced NavSuite. I make these little scratch paper ships to prove to myself that all of the parts work out of the box and not just because of legacy stuff on my production ships. They also serve as a handy example of the parts and how they work together. This one uses the Upward L receiver option that is documented in the Compass_Orient chip. I didn't tune the strafe arrest for such a small ship so its a bit squirmy, but the parts work and just require tuning and customization to a ship's needs.